File moved to Debugging -> CrashHandler .h/cpp
