+++ src/shared/Debugging/CrashHandler.h
